requirements.txt
streamlit==1.32.0
ffmpeg-python==0.2.0
requests==2.31.0
cloudinary==1.38.0
elevenlabs==2.36.0
google-generativeai==0.4.1
python-dotenv==1.0.1
pillow==10.2.0
moviepy==1.0.3
static-ffmpeg==2.5
